// packages/api/src/utils/db-types.ts
export interface definitions {
  account: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    issuer: string
    sub: string
    name: string
    picture?: string
    github?: string
    email: string
    public_address: string
    inserted_at: string
    updated_at: string
  }
  auth_key: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    key_id: number
    name: string
    secret: string
    /**
     * Note:
     * This is a Foreign Key to `account.issuer`.<fk table='account' column='issuer'/>
     */
    issuer: string
    inserted_at: string
    updated_at: string
  }
  content: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    cid: string
    size?: number
    inserted_at: string
    updated_at: string
  }
  pin: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    pin_id: number
    status: 'queued' | 'pinning' | 'pinned' | 'failed'
    /**
     * Note:
     * This is a Foreign Key to `content.cid`.<fk table='content' column='cid'/>
     */
    cid: string
    service: 'PINATA' | 'IPFS_CLUSTER'
    inserted_at: string
    updated_at: string
  }
  upload: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    upload_id: number
    /**
     * Note:
     * This is a Foreign Key to `account.issuer`.<fk table='account' column='issuer'/>
     */
    issuer: string
    /**
     * Note:
     * This is a Foreign Key to `auth_key.key_id`.<fk table='auth_key' column='key_id'/>
     */
    key_id?: number
    /**
     * Note:
     * This is a Foreign Key to `content.cid`.<fk table='content' column='cid'/>
     */
    cid: string
    type: string
    files?: string
    origins?: string
    meta?: string
    name?: string
    inserted_at: string
    updated_at: string
  }
}